+++ CHANGELOG.md
@@ -74,6 +74,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   [#434](https://gitlab.data.bas.ac.uk/MAGIC/add-metadata-toolbox/-/issues/434)
 * Switching to `metadata-resources` site for downloads proxy schema
   [#435](https://gitlab.data.bas.ac.uk/MAGIC/add-metadata-toolbox/-/issues/435)
+* Switching to Embedded Maps Service for extent maps
+  [#256](https://gitlab.data.bas.ac.uk/MAGIC/add-metadata-toolbox/-/issues/256)
 
 ### Removed